Make `ParseDeviceString` just parse
