+++ torch_xla/csrc/ops/max_unpool_nd.cpp
@@ -3,6 +3,7 @@
 #include "torch_xla/csrc/lowering_context.h"
 #include "torch_xla/csrc/ops/infer_output_shape.h"
 #include "torch_xla/csrc/pooling.h"
+#include "torch_xla/csrc/runtime/runtime.h"
 #include "torch_xla/csrc/runtime/debug_macros.h"
 
 namespace torch_xla {
