+++ torch_xla/csrc/runtime/runtime.cc
@@ -1,3 +1,6 @@
+#include <torch/csrc/lazy/backend/backend_device.h>
+
+#include "torch_xla/csrc/device.h"
 #include "torch_xla/csrc/runtime/computation_client.h"
 #include "torch_xla/csrc/runtime/env_vars.h"
 #include "torch_xla/csrc/runtime/pjrt_computation_client.h"
@@ -9,6 +12,7 @@ namespace {
 
 std::atomic<ComputationClient*> g_computation_client(nullptr);
 std::once_flag g_computation_client_once;
+
 
 ComputationClient* CreateClient() {
   if (sys_util::GetEnvBool("XLA_DUMP_FATAL_STACK", false)) {
@@ -41,4 +45,38 @@ ComputationClient* GetComputationClientIfInitialized() {
 }
 
 }  // namespace runtime
+
+namespace {
+
+thread_local absl::optional<torch::lazy::BackendDevice> g_current_device;
+
+}
+
+const torch::lazy::BackendDevice* GetDefaultDevice() {
+  std::string default_device_spec =
+      UseVirtualDevice()
+            ? "SPMD:0"
+            : runtime::GetComputationClient()->GetDefaultDevice();
+  XLA_CHECK(!default_device_spec.empty());
+  static const torch::lazy::BackendDevice default_device =
+      ParseDeviceString(default_device_spec);
+  return &default_device;
+}
+
+torch::lazy::BackendDevice GetCurrentDevice() {
+  if (!g_current_device) {
+    g_current_device = *GetDefaultDevice();
+  }
+  return *g_current_device;
+}
+
+
+torch::lazy::BackendDevice SetCurrentDevice(
+    const torch::lazy::BackendDevice& device) {
+  torch::lazy::BackendDevice current = GetCurrentDevice();
+  g_current_device = device;
+  TF_VLOG(2) << "New current device: " << device;
+  return current;
+}
+
 }  // namespace torch_xla
