+++ torch_xla/csrc/runtime/runtime.h
@@ -1,6 +1,9 @@
 #ifndef XLA_CLIENT_RUNTIME_H_
 #define XLA_CLIENT_RUNTIME_H_
 
+#include <torch/csrc/lazy/backend/backend_device.h>
+
+#include "torch_xla/csrc/device.h"
 #include "torch_xla/csrc/runtime/computation_client.h"
 
 namespace torch_xla {
@@ -16,6 +19,19 @@ ComputationClient* GetComputationClientIfInitialized();
 void RunLocalService(uint64_t service_port);
 
 }  // namespace runtime
+
+const torch::lazy::BackendDevice* GetDefaultDevice();
+
+torch::lazy::BackendDevice GetCurrentDevice();
+
+torch::lazy::BackendDevice SetCurrentDevice(
+    const torch::lazy::BackendDevice& device);
+
+static inline torch::lazy::BackendDevice GetDeviceOrCurrent(
+    const torch::lazy::BackendDevice* device) {
+  return device != nullptr ? *device : GetCurrentDevice();
+}
+
 }  // namespace torch_xla
 
 #endif
